+++ packages/mc-html-template/src/process-headers.spec.ts
@@ -102,8 +102,8 @@ describe('permissionsPolicies', () => {
       ...defaultApplicationConfig,
       headers: {
         permissionsPolicies: {
-          microphone: '()',
-          camera: '(self)',
+          microphone: '*',
+          camera: '()',
         },
       },
     };
@@ -112,25 +112,8 @@ describe('permissionsPolicies', () => {
 
     expect(
       processedApplicationConfig['Permissions-Policy']
-    ).toMatchInlineSnapshot(`"microphone=(), camera=(self)"`);
-  });
-});
-describe('strictTransportSecurity', () => {
-  it('should add extra options', () => {
-    const testApplicationConfig = {
-      ...defaultApplicationConfig,
-      headers: {
-        strictTransportSecurity: ['includeSubDomains'],
-      },
-    };
-
-    // @ts-ignore
-    const processedApplicationConfig = processHeaders(testApplicationConfig);
-
-    expect(
-      processedApplicationConfig['Strict-Transport-Security']
     ).toMatchInlineSnapshot(
-      `"max-age=31536000; includeSubDomains; preload; includeSubDomains"`
+      `"microphone=*, camera=(), payment=(self), usb=(self), geolocation=(self)"`
     );
   });
 });