+++ custom-views-templates/starter-typescript/CHANGELOG.md
@@ -1,5 +1,91 @@
 # @commercetools-applications/merchant-center-custom-view-template-starter-typescript
 
+## 22.38.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-frontend/application-shell@22.38.3
+  - @commercetools-frontend/actions-global@22.38.3
+  - @commercetools-frontend/application-components@22.38.3
+  - @commercetools-frontend/application-config@22.38.3
+  - @commercetools-frontend/application-shell-connectors@22.38.3
+  - @commercetools-frontend/assets@22.38.3
+  - @commercetools-frontend/babel-preset-mc-app@22.38.3
+  - @commercetools-frontend/constants@22.38.3
+  - @commercetools-frontend/eslint-config-mc-app@22.38.3
+  - @commercetools-frontend/i18n@22.38.3
+  - @commercetools-frontend/jest-preset-mc-app@22.38.3
+  - @commercetools-frontend/l10n@22.38.3
+  - @commercetools-frontend/mc-dev-authentication@22.38.3
+  - @commercetools-frontend/mc-scripts@22.38.3
+  - @commercetools-frontend/permissions@22.38.3
+
+## 22.38.2
+
+### Patch Changes
+
+- Updated dependencies [[`f158094`](https://github.com/commercetools/merchant-center-application-kit/commit/f15809419de1cc278484ab07b542eacd99dbf499), [`f158094`](https://github.com/commercetools/merchant-center-application-kit/commit/f15809419de1cc278484ab07b542eacd99dbf499)]:
+  - @commercetools-frontend/constants@22.38.2
+  - @commercetools-frontend/mc-scripts@22.38.2
+  - @commercetools-frontend/actions-global@22.38.2
+  - @commercetools-frontend/application-components@22.38.2
+  - @commercetools-frontend/application-config@22.38.2
+  - @commercetools-frontend/application-shell@22.38.2
+  - @commercetools-frontend/application-shell-connectors@22.38.2
+  - @commercetools-frontend/mc-dev-authentication@22.38.2
+  - @commercetools-frontend/permissions@22.38.2
+  - @commercetools-frontend/i18n@22.38.2
+  - @commercetools-frontend/l10n@22.38.2
+  - @commercetools-frontend/assets@22.38.2
+  - @commercetools-frontend/babel-preset-mc-app@22.38.2
+  - @commercetools-frontend/eslint-config-mc-app@22.38.2
+  - @commercetools-frontend/jest-preset-mc-app@22.38.2
+
+## 22.38.1
+
+### Patch Changes
+
+- Updated dependencies [[`198f30c`](https://github.com/commercetools/merchant-center-application-kit/commit/198f30c30a7c293d08e512070bf1450971100f12), [`2fec608`](https://github.com/commercetools/merchant-center-application-kit/commit/2fec608c3f47de05c83eb7806f89246b6bdc5ae7)]:
+  - @commercetools-frontend/application-components@22.38.1
+  - @commercetools-frontend/application-shell-connectors@22.38.1
+  - @commercetools-frontend/application-shell@22.38.1
+  - @commercetools-frontend/mc-scripts@22.38.1
+  - @commercetools-frontend/permissions@22.38.1
+  - @commercetools-frontend/actions-global@22.38.1
+  - @commercetools-frontend/application-config@22.38.1
+  - @commercetools-frontend/assets@22.38.1
+  - @commercetools-frontend/babel-preset-mc-app@22.38.1
+  - @commercetools-frontend/constants@22.38.1
+  - @commercetools-frontend/eslint-config-mc-app@22.38.1
+  - @commercetools-frontend/i18n@22.38.1
+  - @commercetools-frontend/jest-preset-mc-app@22.38.1
+  - @commercetools-frontend/l10n@22.38.1
+  - @commercetools-frontend/mc-dev-authentication@22.38.1
+
+## 22.38.0
+
+### Patch Changes
+
+- [#3671](https://github.com/commercetools/merchant-center-application-kit/pull/3671) [`a435e7f`](https://github.com/commercetools/merchant-center-application-kit/commit/a435e7fda60f05557b740b27f8c14d3c5ce1060e) Thanks [@renovate](https://github.com/apps/renovate)! - Update UI Kit dependencies to v19.20.1
+
+- Updated dependencies [[`3bcdba5`](https://github.com/commercetools/merchant-center-application-kit/commit/3bcdba5452b8b7b7940b3e8570dbfd6837e4788b), [`455cccf`](https://github.com/commercetools/merchant-center-application-kit/commit/455cccfdbac803fd7f821c6127635c1b2e593f27), [`9504631`](https://github.com/commercetools/merchant-center-application-kit/commit/9504631da837e78cf66c84286d5e7e98c2009148), [`9fde2af`](https://github.com/commercetools/merchant-center-application-kit/commit/9fde2afe40616d728cef32d34d39f682a1bf4e31), [`a435e7f`](https://github.com/commercetools/merchant-center-application-kit/commit/a435e7fda60f05557b740b27f8c14d3c5ce1060e), [`5c33a40`](https://github.com/commercetools/merchant-center-application-kit/commit/5c33a40910d1f46d3d28080f666150fe1d002757), [`9fde2af`](https://github.com/commercetools/merchant-center-application-kit/commit/9fde2afe40616d728cef32d34d39f682a1bf4e31)]:
+  - @commercetools-frontend/application-shell-connectors@22.38.0
+  - @commercetools-frontend/application-config@22.38.0
+  - @commercetools-frontend/mc-scripts@22.38.0
+  - @commercetools-frontend/l10n@22.38.0
+  - @commercetools-frontend/application-components@22.38.0
+  - @commercetools-frontend/application-shell@22.38.0
+  - @commercetools-frontend/permissions@22.38.0
+  - @commercetools-frontend/constants@22.38.0
+  - @commercetools-frontend/actions-global@22.38.0
+  - @commercetools-frontend/i18n@22.38.0
+  - @commercetools-frontend/babel-preset-mc-app@22.38.0
+  - @commercetools-frontend/mc-dev-authentication@22.38.0
+  - @commercetools-frontend/jest-preset-mc-app@22.38.0
+  - @commercetools-frontend/assets@22.38.0
+  - @commercetools-frontend/eslint-config-mc-app@22.38.0
+
 ## 22.37.0
 
 ### Patch Changes