+++ packages/constants/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @commercetools-frontend/constants
 
+## 22.38.3
+
+## 22.38.2
+
+### Patch Changes
+
+- [#3696](https://github.com/commercetools/merchant-center-application-kit/pull/3696) [`f158094`](https://github.com/commercetools/merchant-center-application-kit/commit/f15809419de1cc278484ab07b542eacd99dbf499) Thanks [@emmenko](https://github.com/emmenko)! - Change `Permissions-Policy` default directives to `(self)`.
+
+## 22.38.1
+
+## 22.38.0
+
+### Patch Changes
+
+- [#3688](https://github.com/commercetools/merchant-center-application-kit/pull/3688) [`9fde2af`](https://github.com/commercetools/merchant-center-application-kit/commit/9fde2afe40616d728cef32d34d39f682a1bf4e31) Thanks [@emmenko](https://github.com/emmenko)! - Add `Permissions-Policy` to list of security headers
+
 ## 22.37.0
 
 ## 22.36.0